requirements.txt
Django==5.1.7
djangorestframework==3.15.2
django-filter==25.1
djangorestframework_simplejwt==5.5.0
Markdown==3.7
psycopg2-binary==2.9.10